requirements.txt
crewai~=0.28.8
requests~=2.31.0
langchain~=0.1.16
python-dotenv~=1.0.0